[modules] BREAKING uniformize color handling in data fields.
Currently in sofa there is not rules to describe colors in
the data fields. Some components use Data<Vec3>, other Data<Vec4> and others
are using a Data<string> with private float r,g,b,a member.

So some component accepts color by string "blue" other by passing "0 1 0 1"
and other "0 1 2".

Doing this break the consistency of the user interface (either in XML & python scene as all as
in GUIs like runSofa or qtQuickSofa).

This commit unify all that by using a dedicated type defaulttype::RGBAColor.

The advantages are:
   - expose an unify way to expose color to the user that are "blue", "0 0 0 1", "#CCCCCCAA")
   - a lot of component now support the color defined by string ("blue", "red", ...)
   - factoring of the scanning code that was converting the string into a color (now in defaulttype::RGBAColor instead of duplicated in all components).

I try not to break things,  but as this commit is rather big so I may have broken some stuff.
Please test it on one or two scene of your choice and report problem.