[SofaPython] FIX test failure because of sys.argv.
The problem was that runFile is used either in the SceneLoader and to load
a controller but the behavior should be different regarding the way
the sys.argv have been set.  The solution was to revert (with small refactoring)
to the initial runFile behavior but adding a call to setArgument in the python
SceneLoader.