[SofaPython] Base.addNewData returns the created data or set an exception + corresponding tests.
The addNewData allows to create a new data field to sofa object or node.
It is better to returns the created object instead of None so we can immediately
start using it without going through a costly lookup in the field's map.

The function was also raising an exception in case no data were created but
the string associated to the exception was not set.

This commit change that.